+++ providers/dns/loopia/loopia.go
@@ -6,11 +6,15 @@ import (
 	"fmt"
 	"net/http"
 	"strings"
+	"sync"
 	"time"
 
 	"github.com/go-acme/lego/v4/challenge/dns01"
 	"github.com/go-acme/lego/v4/platform/config/env"
+	"github.com/go-acme/lego/v4/providers/dns/loopia/internal"
 )
+
+const minTTL = 300
 
 // Environment variables names.
 const (
@@ -30,7 +34,6 @@ type Config struct {
 	BaseURL            string
 	APIUser            string
 	APIPassword        string
-	Client             *dnsClient
 	PropagationTimeout time.Duration
 	PollingInterval    time.Duration
 	TTL                int
@@ -40,7 +43,7 @@ type Config struct {
 // NewDefaultConfig returns a default configuration for the DNSProvider.
 func NewDefaultConfig() *Config {
 	return &Config{
-		BaseURL:            defaultBaseURL,
+		BaseURL:            internal.DefaultBaseURL,
 		TTL:                env.GetOrDefaultInt(EnvTTL, minTTL),
 		PropagationTimeout: env.GetOrDefaultSecond(EnvPropagationTimeout, 40*time.Minute),
 		PollingInterval:    env.GetOrDefaultSecond(EnvPollingInterval, 60*time.Second),
@@ -52,21 +55,25 @@ func NewDefaultConfig() *Config {
 
 // DNSProvider implements the challenge.Provider interface.
 type DNSProvider struct {
-	config         *Config
-	client         dnsClient
-	findZoneByFqdn func(fqdn string) (string, error)
+	config *Config
+	client *internal.Client
+
 	inProgressInfo map[string]int
+	inProgressMu   sync.Mutex
+
+	findZoneByFqdn func(fqdn string) (string, error)
 }
 
 // NewDNSProvider returns a DNSProvider instance configured for Loopia.
-// Credentials must be passed in the environment variables LOOPIA_API_USER and LOOPIA_API_PASSWORD.
+// Credentials must be passed in the environment variables:
+// LOOPIA_API_USER, LOOPIA_API_PASSWORD.
 func NewDNSProvider() (*DNSProvider, error) {
 	values, err := env.Get(EnvAPIUser, EnvAPIPassword)
 	if err != nil {
 		return nil, fmt.Errorf("loopia: %w", err)
 	}
-	config := NewDefaultConfig()
 
+	config := NewDefaultConfig()
 	config.APIUser = values[EnvAPIUser]
 	config.APIPassword = values[EnvAPIPassword]
 
@@ -82,73 +89,96 @@ func NewDNSProviderConfig(config *Config) (*DNSProvider, error) {
 	if config.APIUser == "" || config.APIPassword == "" {
 		return nil, errors.New("loopia: credentials missing")
 	}
+
 	// Min value for TTL is 300
 	if config.TTL < 300 {
 		config.TTL = 300
 	}
 
-	client := NewClient(config.APIUser, config.APIPassword)
+	client := internal.NewClient(config.APIUser, config.APIPassword)
+
+	// FIXME
 	client.HTTPClient = config.HTTPClient
 
 	return &DNSProvider{
 		config:         config,
-		client:         &client,
+		client:         client,
 		findZoneByFqdn: dns01.FindZoneByFqdn,
 		inProgressInfo: make(map[string]int),
 	}, nil
 }
 
+// Timeout returns the timeout and interval to use when checking for DNS propagation.
+// Adjusting here to cope with spikes in propagation times.
+func (d *DNSProvider) Timeout() (timeout, interval time.Duration) {
+	return d.config.PropagationTimeout, d.config.PollingInterval
+}
+
 // Present creates a TXT record using the specified parameters.
 func (d *DNSProvider) Present(domain, token, keyAuth string) error {
 	fqdn, value := dns01.GetRecord(domain, keyAuth)
+
 	subdomain, authZone := d.splitDomain(fqdn)
-	err := d.client.addTXTRecord(authZone, subdomain, d.config.TTL, value)
+
+	err := d.client.AddTXTRecord(authZone, subdomain, d.config.TTL, value)
 	if err != nil {
-		return err
+		return fmt.Errorf("loopia: %w", err)
 	}
-	txtRecords, err := d.client.getTXTRecords(authZone, subdomain)
+
+	txtRecords, err := d.client.GetTXTRecords(authZone, subdomain)
 	if err != nil {
-		return err
+		return fmt.Errorf("loopia: %w", err)
 	}
+
+	d.inProgressMu.Lock()
+	defer d.inProgressMu.Unlock()
+
 	for _, r := range txtRecords {
 		if r.Rdata == value {
 			d.inProgressInfo[token] = r.RecordID
 			return nil
 		}
 	}
-	return fmt.Errorf("loopia: Failed to get id for TXT record")
+
+	return errors.New("loopia: Failed to get id for TXT record")
 }
 
-// CleanUp removes the TXT record matching the specified
-// parameters.
+// CleanUp removes the TXT record matching the specified parameters.
 func (d *DNSProvider) CleanUp(domain, token, keyAuth string) error {
 	fqdn, _ := dns01.GetRecord(domain, keyAuth)
+
 	subdomain, authZone := d.splitDomain(fqdn)
-	err := d.client.removeTXTRecord(authZone, subdomain, d.inProgressInfo[token])
+
+	d.inProgressMu.Lock()
+	defer d.inProgressMu.Unlock()
+
+	err := d.client.RemoveTXTRecord(authZone, subdomain, d.inProgressInfo[token])
 	if err != nil {
-		return err
+		return fmt.Errorf("loopia: %w", err)
 	}
-	records, err := d.client.getTXTRecords(authZone, subdomain)
+
+	records, err := d.client.GetTXTRecords(authZone, subdomain)
 	if err != nil {
-		return err
+		return fmt.Errorf("loopia: %w", err)
 	}
-	if len(records) == 0 {
-		err = d.client.removeSubdomain(authZone, subdomain)
+
+	if len(records) > 0 {
+		return nil
 	}
-	return err
+
+	err = d.client.RemoveSubdomain(authZone, subdomain)
+	if err != nil {
+		return fmt.Errorf("loopia: %w", err)
+	}
+
+	return nil
 }
 
 func (d *DNSProvider) splitDomain(fqdn string) (string, string) {
 	authZone, _ := d.findZoneByFqdn(fqdn)
 	authZone = dns01.UnFqdn(authZone)
-	unFqdn := dns01.UnFqdn(fqdn)
-	subdomain := strings.TrimSuffix(unFqdn, "."+authZone)
-	return subdomain, authZone
-}
 
-// Timeout returns the values (40*time.Minute, 60*time.Second) which
-// are used by the acme package as timeout and check interval values
-// when checking for DNS record propagation with Loopia.
-func (d *DNSProvider) Timeout() (timeout, interval time.Duration) {
-	return d.config.PropagationTimeout, d.config.PollingInterval
+	subdomain := strings.TrimSuffix(dns01.UnFqdn(fqdn), "."+authZone)
+
+	return subdomain, authZone
 }
